reader_py/requirements.txt
timex-linux-global-trainer-gps
lxml>=3.3.3
